RETURNING excludes computed cols
WIth POSTGRES supporting computed/generated columns, it'd be great to have those columns come back in the RETURNING clause, instead of needing to reload the values after a mutation. In `schema/schema.go` around line 182, it looks like `FieldsWithDefaultDBValue` property on the schema is what's used to determine what should be generated in the RETURNING clause.