Show difference between subquery types
TestSubQueryRaw shows that an error occurs, and unexpected SQL,
when a sub query is made using statement built using the .Raw
method.

TestSubQueryBuilder shows the equivalent expected behavior.

In this simple example there would be no reason not to just use
the normal GORM builder, or just a String for the subquery. In
more complex usecases we still would like to use GORM's API
in order to leverage params/args, and general code reuse. Raw
is used to use the UNION operator, for example.