Bug: self-referential has-one does not create foreign key
In the existing models of this repo the user foreign key (manager_id)
gets created by the has-many relation. When that is no longer present no
foreign key is created for the self-referential has-one relation
corresponding to the Manager member.

This is a test case which shows this bug with a separate model.