Create test case demonstrating constraint name bug
This commit triggers a bug where an invalid foreign key constraint name
is generated due to the length of the table name and the fact that the
table name includes a non-default schema.

The invalid constraint looks like this:

```
CONSTRAINT "fkmy_schema"."table_oneeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeta520923f2"
```

But it should look like:

```
CONSTRAINT "fkmy_schema_table_oneeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeta520923f2"
```

If the table name is any shorter, or if the table doesn't include a
schema name, then the bug is not triggered.