test: reproduce deadlock in `prepare`
Deadlocks as written. All connections in the pool are acquired by
clients that are waiting for a statement to be prepared, but the caller
in charge of preparing that statement needs a connection to do so.

Passes if you change `pool_max_conns` from 2 to 3 or higher, or if you
set `PrepareStmt: false`.